requirements.txt
aiosqlite
annotated-types
anyio
blinker
certifi
click
colorama
Flask
greenlet
h11
httpcore
httpx
httpx-sse
idna
itsdangerous
Jinja2
markdown-it-py
MarkupSafe
mcp
mdurl
peewee
pydantic
pydantic-settings
pydantic_core
Pygments
python-dotenv
python-multipart
rich
shellingham
sniffio
SQLAlchemy
sqlite-web
sse-starlette
starlette
typer
typing-inspection
typing_extensions
uvicorn
Werkzeug
fastapi
